+++ client/src/components/existing-request/existing-request-display.vue
@@ -164,7 +164,6 @@ import newReqModule from '@/store/new-request-module'
 import NrAffiliationMixin from '@/components/mixins/nr-affiliation-mixin'
 import PaymentMixin from '@/components/payment/payment-mixin'
 import CommonMixin from '@/components/mixins/common-mixin'
-import OutputMixin from '@/components/mixins/output-mixin'
 import DateMixin from '@/components/mixins/date-mixin'
 import paymentModule from '@/modules/payment'
 import timerModule from '@/modules/vx-timer'
@@ -194,7 +193,6 @@ export default class ExistingRequestDisplay extends Mixins(
   NrAffiliationMixin,
   CommonMixin,
   DateMixin,
-  OutputMixin,
   PaymentMixin) {
   // enums used in the template:
   NameState = NameState
@@ -531,7 +529,14 @@ export default class ExistingRequestDisplay extends Mixins(
           this.navigateToPaymentPortal()
           break
         case NrAction.RESULTS:
-          await this.downloadOutputs(this.nr.id)
+          // show spinner since the network calls below can take a few seconds
+          this.$root.$emit('showSpinner', true)
+
+          // Request the outputs
+          await newReqModule.downloadOutputs(this.nr.id)
+
+          // clear spinner
+          this.$root.$emit('showSpinner', false)
           break
         default:
           if (await newReqModule.patchNameRequestsByAction(action)) {
